chatbot-tenders/backend/requirements.txt
fastapi
uvicorn[standard]
elasticsearch
psycopg2-binary
sqlalchemy
pydantic
python-multipart
nltk
spacy
transformers
torch
python-dotenv
httpx
pandas
numpy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
asyncpg